+++ RELEASENOTES.md
@@ -1,5 +1,7 @@
 These are the release notes for the [wikibase-api](README.md).
 
+## Version
+
 ## Version 0.5 (2015-06-29)
 
 ####Additions